+++ rdkit/Chem/Graphs.py
@@ -1,6 +1,5 @@
-# $Id$
 #
-# Copyright (C) 2001-2008 greg landrum and rational discovery llc
+# Copyright (C) 2001-2022 Greg Landrum and other RDKit contributors
 #
 #   @@ All Rights Reserved @@
 #  This file is part of the RDKit.
@@ -14,14 +13,15 @@ In theory much of the functionality in here should be migrating into the
 C/C++ codebase.
 
 """
-import numpy
-from rdkit import Chem
-from rdkit import DataStructs
 import types
+
+import numpy
+
+from rdkit import Chem, DataStructs
 
 
 def CharacteristicPolynomial(mol, mat=None):
-    """ calculates the characteristic polynomial for a molecular graph
+  """ calculates the characteristic polynomial for a molecular graph
 
       if mat is not passed in, the molecule's Weighted Adjacency Matrix will
       be used.
@@ -31,21 +31,21 @@ def CharacteristicPolynomial(mol, mat=None):
       1992), pg 76.
 
     """
-    nAtoms = mol.GetNumAtoms()
-    if mat is None:
-        # FIX: complete this:
-        #A = mol.GetWeightedAdjacencyMatrix()
-        pass
-    else:
-        A = mat
-    I = 1. * numpy.identity(nAtoms)
-    An = A
-    res = numpy.zeros(nAtoms + 1, numpy.float)
-    res[0] = 1.0
-    for n in range(1, nAtoms + 1):
-        res[n] = 1. / n * numpy.trace(An)
-        Bn = An - res[n] * I
-        An = numpy.dot(A, Bn)
+  nAtoms = mol.GetNumAtoms()
+  if mat is None:
+    # FIX: complete this:
+    #A = mol.GetWeightedAdjacencyMatrix()
+    pass
+  else:
+    A = mat
+  I = 1. * numpy.identity(nAtoms)
+  An = A
+  res = numpy.zeros(nAtoms + 1, float)
+  res[0] = 1.0
+  for n in range(1, nAtoms + 1):
+    res[n] = 1. / n * numpy.trace(An)
+    Bn = An - res[n] * I
+    An = numpy.dot(A, Bn)
 
-    res[1:] *= -1
-    return res
+  res[1:] *= -1
+  return res
