+++ rdkit/ML/InfoTheory/BitRank.py
@@ -78,7 +78,7 @@ def CalcInfoGains(bitVects, actVals, nPossibleActs, nPossibleBitVals=2):
   if len(bitVects) != len(actVals):
     raise ValueError('var and activity lists should be the same length')
   nBits = len(bitVects[0])
-  res = numpy.zeros(nBits, numpy.float)
+  res = numpy.zeros(nBits, float)
 
   for bit in range(nBits):
     counts = FormCounts(bitVects, actVals, bit, nPossibleActs, nPossibleBitVals=nPossibleBitVals)
